Added Gobelins, l'école de l'image
School of visual communication located in Paris with main focus on 2D animation with some courses on 3D and video games.
In this school, IT is mainly used for interactive design (http://www.gobelins.fr/formations/initiale/design-interactif-web FR), with course duration going from summer school to 4 years. The Creative Lead / Creative Tech course takes 2 years to complete and is based on web and mobile technologies.