Added Cégep de Drummondville as a school
Their website: http://www.cegepdrummond.ca/ (the www. is important)
The IT program (3 years):

http://www.cegepdrummond.ca/futurs-etudiants/nos-programmes/dec-techniques/techniques-de-linformatique-specialisation-en-informatique-de-gestion/
The website is in the same domain as the email domain, so it should not
be a problem. The format of the email is: studentId@cegepdrummond.ca

Sorry, as it is a french college (Quebec, Canada), it is only available
in french.