Add files via upload
http://cnstefancelmare.ro/?pagina=6
Here are listed the courses "Mate-Info" or "Matematica-Informatica" is Maths and Informatics.

https://imgur.com/a/MQZaA72

And here is a screenshot from the email that a CS teacher in my school sent guiding us in using the microsoft office platform for online learning. If this isn't enough i can send the pdf as well.