Create mail-ecv.txt
mail-ecv.fr is a new mail extension for ecv (as ecv digital) schools. My school page is located there : https://www.ecv.fr/lecole/les-campus/bordeaux/
You can find here a proof that this e-mail domain is used for the school : https://i.ibb.co/pKSJwLF/image.png
Moreover, the domain of this page is https://extranet.campus-ade.com. If you go at this page, you can see the logo of ecv.
If it's not enough, I'll ask the administration of the school an official document saying that this e-mail is linked to ecv.