feat: Added school
BBZ Berufsbildungszentrum Rendsburg-Eckernförde

For easier translation: This is a german school.
It offers both higher education and apprenticeships.

URL to the homepage:
https://www.bbz-rd-eck.de/

URL to the IT education
https://www.bbz-rd-eck.de/bildungsangebote/duale-ausbildung/it-berufe/
(Fachinformatiker für Anwendungsentwicklung = Application development)

Proof of the E-Mail for the top-level:
https://www.bbz-rd-eck.de/online-sekretariat/schulbuero/
(Kontaktdaten = Contact information)

I can't really find a proof for the "sus" (Schülerinnen und Schüler = Pupils). But it's used to seperate students from staff or teachers.