Add files via upload
I applied to add my school to jet brains in order to get jet brains' free products and a school website URL was requested but my school doesn't have one. So all I can provide is the school's address and a document that proves that I am a student in the school.