Add darzo.txt
Add the high school(istituto superiore) Silvio D'arzo to the list of accepted institutions.

Its official website is: https://www.istitutodarzo.edu.it/
The school offers many educational programs(or courses), among which is a CS-related one, which lasts five years. A short description along with a syllabus of the aforementioned can be found here: https://www.istitutodarzo.edu.it/informatica-e-telecomunicazioni/
Unfortunately I couldn't find anywhere in the website something stating that each student will be provided with a scholastic email address, so I'm afraid a screenshot of mine is about as much evidence as I can present. 
![evidence](https://drive.google.com/file/d/1oXyUoGIXaMTA_fWh6gqpDuywrcPKrnOH/view?usp=sharing)