Added new University
Hello. I would like to attach the institutional website (https://ugv.edu.br/), (https://uniao.ugv.edu.br/). I had sent this request once before, but an error occurred when entering the website, so I would like to ask that if it happens again, if it is possible to search for it on the web (Ugv) and the result will appear immediately. The university is located in the state of Paraná, Brazil in the city União da Vitória.