Add files via upload
This is the Govt. Institution that provides training to all the teachers of School Education Department of Government of Punjab, Pakistan. There are many computer programming courses that are offered. 
Here is the url of School Education Department: https://schools.punjab.gov.pk/

Here is the URL where academics and trainings offered for teaching staff:
http://qaed.edu.pk/pages/subpages/6