Add files via upload
Recommit- Mistakenly added to us folder instead of edu folder. Removed rvu.txt from us and added it to edu.
RVU.edu (Rocky Vista University) There is a Colorado and a Utah campus so emails will be ut.rvu.edu or co.rvu.edu. This is the main website
https://www.rvu.edu/

This is the catalog showing the Digital Health Track which is a class for which these materials are actually useful (considering this is a medical school a good amount of the classes would have no need for coding, however, the DHT is a 4-year program as described on the site where integration of digital health, ai, and other technology is used to try and advance medical education and logistics.)

https://catalog.rvu.edu/dgt-4001